requirements.txt
python-telegram-bot==13.15
python-dotenv==0.20
pandas==1.5.2
openpyxl==3.0.10
xlrd==2.0.1